requirements.txt
argon2-cffi==20.1.0
bcrypt==3.2.0
brotlipy==0.7.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
comtypes==1.1.8
cryptography==3.3.1
Django==2.2.5
django-crispy-forms==1.10.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
et-xmlfile==1.0.1
jdcal==1.4.1
pdfminer==20191125
Pillow==8.0.1
pycparser==2.20
pycryptodome==3.9.9
PyJWT==1.7.1
PySocks==1.7.1
python-dateutil==2.8.1
pywin32==227
requests==2.25.1
six==1.15.0
somewhere==0.0.1
sqlparse==0.3.0
tqdm==4.56.2
urllib3==1.26.3
win-inet-pton==1.1.0
wincertstore==0.2
xlrd==1.2.0
xlutils==2.0.0
xlwt==1.3.0
